Add support for python 3.10
Python 3.10 moved the Sequence type under collections.abc, rendering it
incompatible with previous versions. Try to import with this new location
and fallback to previous location to keep backward compatibility with
Python2.

setuptools-mardown is now deprecated, in favor of the
long_description_content_type argument in setup.py. This replacement is
done because the setuptools-mardown library depends on any version of
pypandoc and is now broken with the latest version (1.8) which removed
a function that was used, causing this error:
AttributeError: module 'pypandoc' has no attribute 'convert'

In setup.py, the setup_require key is now deprecated, and has been removed
since build build only needed setuptools-mardown which is no longer needed.
In setup.py, the test_suite and tests_require keys are now deprecated,
in favor of an external testing tool. The tox configuration file now
contain the description of the required test dependencies.

Nose is replaced by pytest in this commit, mainly because I was unable to
run any test using it, and pytest worked as a drop-in replacement.